slug folder ის შეცვლა გადაწერა server side ზე
