+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.7.7",
     "mysql2": "^3.11.3",
     "next": "14.2.15",
     "react": "^18",
